+++ src/main/java/com/raoulvdberge/refinedstorageaddons/RSAddonsGui.java
@@ -1,5 +1,5 @@
 package com.raoulvdberge.refinedstorageaddons;
 
-public class RSAddonsGui {
+public final class RSAddonsGui {
     public static final int INFINITE_WIRELESS_TRANSMITTER = 0;
 }