+++ src/main/java/com/raoulvdberge/refinedstorageaddons/gui/GuiHandler.java
@@ -4,6 +4,7 @@ import com.raoulvdberge.refinedstorageaddons.RSAddonsGui;
 import com.raoulvdberge.refinedstorageaddons.container.ContainerInfiniteWirelessTransmitter;
 import com.raoulvdberge.refinedstorageaddons.tile.TileInfiniteWirelessTransmitter;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.common.network.IGuiHandler;
@@ -14,9 +15,10 @@ public class GuiHandler implements IGuiHandler {
     @Nullable
     @Override
     public Object getServerGuiElement(int ID, EntityPlayer player, World world, int x, int y, int z) {
+        TileEntity tile = world.getTileEntity(new BlockPos(x, y, z));
         switch (ID) {
             case RSAddonsGui.INFINITE_WIRELESS_TRANSMITTER:
-                return new ContainerInfiniteWirelessTransmitter((TileInfiniteWirelessTransmitter) world.getTileEntity(new BlockPos(x, y, z)), player);
+                return new ContainerInfiniteWirelessTransmitter((TileInfiniteWirelessTransmitter) tile, player);
         }
         return null;
     }