+++ src/main/java/com/raoulvdberge/refinedstorageaddons/proxy/ProxyCommon.java
@@ -7,13 +7,16 @@ import com.raoulvdberge.refinedstorageaddons.RSAddonsItems;
 import com.raoulvdberge.refinedstorageaddons.apiimpl.network.grid.wireless.WirelessGridFactoryWirelessCraftingGrid;
 import com.raoulvdberge.refinedstorageaddons.apiimpl.network.node.NetworkNodeInfiniteWirelessTransmitter;
 import com.raoulvdberge.refinedstorageaddons.gui.GuiHandler;
+import com.raoulvdberge.refinedstorageaddons.item.ItemNetworkBag;
 import com.raoulvdberge.refinedstorageaddons.item.WirelessCraftingGrid;
 import com.raoulvdberge.refinedstorageaddons.network.MessagePickBlock;
 import com.raoulvdberge.refinedstorageaddons.tile.TileInfiniteWirelessTransmitter;
 import net.minecraft.block.Block;
 import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.event.entity.player.EntityItemPickupEvent;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
@@ -50,6 +53,21 @@ public class ProxyCommon {
     }
 
     @SubscribeEvent
+    public void onPickupItem(EntityItemPickupEvent e) {
+        if (!e.getEntityPlayer().world.isRemote) {
+            for (int i = 0; i < 9 * 4; ++i) {
+                ItemStack stack = e.getEntityPlayer().inventory.getStackInSlot(i);
+
+                if (stack.getItem() == RSAddonsItems.NETWORK_BAG) {
+                    ((ItemNetworkBag) stack.getItem()).onPlayerPickup(e, stack);
+
+                    break;
+                }
+            }
+        }
+    }
+
+    @SubscribeEvent
     public void registerBlocks(RegistryEvent.Register<Block> e) {
         e.getRegistry().register(RSAddonsBlocks.INFINITE_WIRELESS_TRANSMITTER);
     }
@@ -58,6 +76,8 @@ public class ProxyCommon {
     public void registerItems(RegistryEvent.Register<Item> e) {
         e.getRegistry().register(RSAddonsItems.WIRELESS_CRAFTING_GRID);
         e.getRegistry().register(RSAddonsItems.NETWORK_PICKER);
+        e.getRegistry().register(RSAddonsItems.NETWORK_BAG);
+
         e.getRegistry().register(RSAddonsBlocks.INFINITE_WIRELESS_TRANSMITTER.createItem());
     }
 }